# Lazy loading of UnicodeRegex attributes for faster import