UnicodeRegex attributes were computed at import time which added ~1 second overhead to the execution of `import sacrebleu` in downstream packages.

I made them load only when needed using functools.lru_cache. There might be a less convoluted way to do this, but this is the best I could think of for now.
This decreases the import time from 1.6s to 0.2s on my computer.

All tests pass with `pytest test`.

### Before
```
$ echo "import sacrebleu" > /tmp/import_sacrebleu.py; python -m cProfile -s cumtime /tmp/import_sacrebleu.py | head -n 50
         6811583 function calls (6810768 primitive calls) in 1.744 seconds

   Ordered by: cumulative time

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
     63/1    0.005    0.000    1.744    1.744 {built-in method builtins.exec}
        1    0.000    0.000    1.744    1.744 import_sacrebleu.py:1(<module>)
     71/1    0.001    0.000    1.744    1.744 <frozen importlib._bootstrap>:966(_find_and_load)
     71/1    0.000    0.000    1.744    1.744 <frozen importlib._bootstrap>:936(_find_and_load_unlocked)
     71/1    0.001    0.000    1.718    1.718 <frozen importlib._bootstrap>:651(_load_unlocked)
     52/1    0.000    0.000    1.718    1.718 <frozen importlib._bootstrap_external>:672(exec_module)
     95/1    0.000    0.000    1.717    1.717 <frozen importlib._bootstrap>:211(_call_with_frames_removed)
        1    0.000    0.000    1.717    1.717 sacrebleu.py:23(<module>)
  321/319    0.007    0.000    1.559    0.005 {built-in method builtins.__build_class__}
        1    0.000    0.000    1.530    1.530 sacrebleu.py:653(UnicodeRegex)
     1129    0.001    0.000    1.511    0.001 {method 'join' of 'str' objects}
        2    0.000    0.000    1.510    0.755 sacrebleu.py:657(_property_chars)
     7649    0.784    0.000    1.509    0.000 sacrebleu.py:658(<genexpr>)
  2228222    0.259    0.000    0.259    0.000 {built-in method unicodedata.category}
  2231668    0.251    0.000    0.251    0.000 {method 'startswith' of 'str' objects}
  2236033    0.216    0.000    0.216    0.000 {built-in method builtins.chr}
        1    0.000    0.000    0.106    0.106 request.py:68(<module>)
        1    0.000    0.000    0.068    0.068 client.py:69(<module>)
       52    0.001    0.000    0.054    0.001 <frozen importlib._bootstrap_external>:743(get_code)
        1    0.000    0.000    0.050    0.050 parser.py:5(<module>)
        1    0.000    0.000    0.049    0.049 feedparser.py:20(<module>)
       71    0.000    0.000    0.048    0.001 <frozen importlib._bootstrap>:564(module_from_spec)
        1    0.000    0.000    0.046    0.046 _policybase.py:4(<module>)
       16    0.000    0.000    0.045    0.003 <frozen importlib._bootstrap_external>:919(create_module)
       16    0.045    0.003    0.045    0.003 {built-in method _imp.create_dynamic}
       71    0.001    0.000    0.040    0.001 <frozen importlib._bootstrap>:870(_find_spec)
       68    0.000    0.000    0.038    0.001 <frozen importlib._bootstrap_external>:1149(find_spec)
       68    0.001    0.000    0.038    0.001 <frozen importlib._bootstrap_external>:1117(_get_spec)
       52    0.036    0.001    0.037    0.001 <frozen importlib._bootstrap_external>:830(get_data)
      184    0.003    0.000    0.036    0.000 <frozen importlib._bootstrap_external>:1233(find_spec)
        1    0.000    0.000    0.035    0.035 utils.py:5(<module>)
       26    0.000    0.000    0.033    0.001 re.py:231(compile)
       27    0.001    0.000    0.032    0.001 re.py:286(_compile)
       25    0.000    0.000    0.032    0.001 sre_compile.py:557(compile)
        4    0.005    0.001    0.026    0.006 <frozen importlib._bootstrap_external>:1281(_fill_cache)
        1    0.000    0.000    0.025    0.025 hashlib.py:54(<module>)
       14    0.000    0.000    0.023    0.002 hashlib.py:116(__get_openssl_constructor)
        8    0.000    0.000    0.023    0.003 hashlib.py:73(__get_builtin_constructor)
        4    0.021    0.005    0.021    0.005 {built-in method posix.listdir}
       25    0.000    0.000    0.020    0.001 sre_parse.py:844(parse)
    59/25    0.000    0.000    0.019    0.001 sre_parse.py:407(_parse_sub)
    72/28    0.008    0.000    0.019    0.001 sre_parse.py:470(_parse)
        1    0.000    0.000    0.018    0.018 __init__.py:24(<module>)
        1    0.000    0.000    0.017    0.017 typing.py:1(<module>)
      5/4    0.000    0.000    0.016    0.004 {built-in method builtins.__import__}
```

### After
```
$ echo "import sacrebleu" > /tmp/import_sacrebleu.py; python -m cProfile -s cumtime /tmp/import_sacrebleu.py | head -n 50
         55679 function calls (54877 primitive calls) in 0.182 seconds

   Ordered by: cumulative time

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
     63/1    0.007    0.000    0.182    0.182 {built-in method builtins.exec}
        1    0.000    0.000    0.182    0.182 import_sacrebleu.py:1(<module>)
     71/1    0.001    0.000    0.182    0.182 <frozen importlib._bootstrap>:966(_find_and_load)
     71/1    0.000    0.000    0.182    0.182 <frozen importlib._bootstrap>:936(_find_and_load_unlocked)
     71/1    0.001    0.000    0.155    0.155 <frozen importlib._bootstrap>:651(_load_unlocked)
     52/1    0.000    0.000    0.155    0.155 <frozen importlib._bootstrap_external>:672(exec_module)
     95/1    0.000    0.000    0.154    0.154 <frozen importlib._bootstrap>:211(_call_with_frames_removed)
        1    0.000    0.000    0.154    0.154 sacrebleu.py:23(<module>)
        1    0.000    0.000    0.091    0.091 request.py:68(<module>)
        1    0.000    0.000    0.061    0.061 client.py:69(<module>)
       71    0.001    0.000    0.048    0.001 <frozen importlib._bootstrap>:870(_find_spec)
       68    0.000    0.000    0.046    0.001 <frozen importlib._bootstrap_external>:1149(find_spec)
       68    0.001    0.000    0.046    0.001 <frozen importlib._bootstrap_external>:1117(_get_spec)
      181    0.003    0.000    0.045    0.000 <frozen importlib._bootstrap_external>:1233(find_spec)
        1    0.000    0.000    0.042    0.042 parser.py:5(<module>)
        1    0.000    0.000    0.041    0.041 feedparser.py:20(<module>)
        1    0.000    0.000    0.037    0.037 _policybase.py:4(<module>)
       52    0.001    0.000    0.037    0.001 <frozen importlib._bootstrap_external>:743(get_code)
  322/320    0.010    0.000    0.033    0.000 {built-in method builtins.__build_class__}
        4    0.005    0.001    0.028    0.007 <frozen importlib._bootstrap_external>:1281(_fill_cache)
        1    0.000    0.000    0.026    0.026 utils.py:5(<module>)
        4    0.023    0.006    0.023    0.006 {built-in method posix.listdir}
        1    0.000    0.000    0.021    0.021 typing.py:1(<module>)
       52    0.018    0.000    0.019    0.000 <frozen importlib._bootstrap_external>:830(get_data)
       71    0.000    0.000    0.017    0.000 <frozen importlib._bootstrap>:564(module_from_spec)
      5/4    0.000    0.000    0.017    0.004 {built-in method builtins.__import__}
       85    0.002    0.000    0.016    0.000 typing.py:922(__new__)
       16    0.000    0.000    0.014    0.001 <frozen importlib._bootstrap_external>:919(create_module)
        1    0.000    0.000    0.014    0.014 __init__.py:24(<module>)
       16    0.014    0.001    0.014    0.001 {built-in method _imp.create_dynamic}
       52    0.000    0.000    0.013    0.000 <frozen importlib._bootstrap_external>:485(_compile_bytecode)
       23    0.000    0.000    0.013    0.001 re.py:231(compile)
       24    0.000    0.000    0.013    0.001 re.py:286(_compile)
       52    0.013    0.000    0.013    0.000 {built-in method marshal.loads}
       22    0.000    0.000    0.012    0.001 sre_compile.py:557(compile)
  131/120    0.000    0.000    0.012    0.000 <frozen importlib._bootstrap>:997(_handle_fromlist)
      101    0.002    0.000    0.011    0.000 abc.py:132(__new__)
        1    0.000    0.000    0.011    0.011 tempfile.py:24(<module>)
       95    0.000    0.000    0.010    0.000 typing.py:133(__new__)
      317    0.000    0.000    0.010    0.000 <frozen importlib._bootstrap_external>:75(_path_stat)
        1    0.000    0.000    0.010    0.010 ssl.py:91(<module>)
      317    0.010    0.000    0.010    0.000 {built-in method posix.stat}
    71/68    0.000    0.000    0.010    0.000 typing.py:676(inner)
        1    0.000    0.000    0.009    0.009 shutil.py:5(<module>)
       39    0.001    0.000    0.009    0.000 typing.py:1099(__getitem__)
```