@mjpost 2019-07-12
This is really great, thank you! I'll bundle up 1.3.7 and release it soon.